# api/openapi.yaml
openapi: 3.0.3
info: {title: Netlab API, version: 1.0.0}
servers: [{url: /api/v1}]
security: [{bearer: []}]
paths:
  /templates/{id}/cache:
    parameters: [{$ref: '#/components/parameters/Id'}]
    get:
      operationId: templateCache
      responses:
        '200': {description: Local base disk cache, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/TemplateCache'}}}}}
    delete:
      operationId: trimTemplateCache
      responses: {'202': {$ref: '#/components/responses/Accepted'}}
...
  /nodes/{id}/storage-device:
    parameters: [{$ref: '#/components/parameters/Id'}]
    get:
      operationId: nodeStorageDevices
      responses:
        '200': {description: Physical disks, selected Ceph disk and preparation task, content: {application/json: {schema: {$ref: '#/components/schemas/NodeStorageDevices'}}}}
    put:
      operationId: configureNodeStorage
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ConfigureNodeStorage'}}}}
      responses:
        '202': {$ref: '#/components/responses/Accepted'}
  /storage-pools/{id}/ceph:
    parameters: [{$ref: '#/components/parameters/Id'}]
    get:
      operationId: cephStatus
      responses:
        '200': {description: Live managed Ceph status, content: {application/json: {schema: {$ref: '#/components/schemas/CephStatus'}}}}
    put:
      operationId: configureCephPool
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ConfigureCephPool'}}}}
      responses:
        '202': {$ref: '#/components/responses/Accepted'}
  /storage-pools/{id}/assets:
    parameters: [{$ref: '#/components/parameters/Id'}]
    get:
      operationId: storagePoolAssets
      responses:
        '200': {description: Assets using this pool, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/StoragePoolAsset'}}}}}
...
  /storage-pools:
    get:
      operationId: listStoragePools
      responses:
        '200': {description: Storage pools and current capacity, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/StoragePool'}}}}}
    post:
      operationId: createStoragePool
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CreateStoragePool'}}}}
      responses:
        '201': {description: Registered storage, content: {application/json: {schema: {$ref: '#/components/schemas/StoragePool'}}}}
  /storage-pools/{id}:
    parameters: [{$ref: '#/components/parameters/Id'}]
    delete:
      operationId: deleteStoragePool
      responses:
        '202': {$ref: '#/components/responses/Accepted'}
        '409': {description: Pool is referenced, content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}}
...
  /operations:
    get:
      operationId: listOperations
      parameters: [{name: environmentId, in: query, schema: {type: string}}, {$ref: '#/components/parameters/Cursor'}, {$ref: '#/components/parameters/Limit'}]
      responses:
        '200': {description: Operations, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/Operation'}}}}}
  /operations/{id}:
    parameters: [{$ref: '#/components/parameters/Id'}]
    get:
      operationId: getOperation
      responses:
        '200': {description: Operation, content: {application/json: {schema: {$ref: '#/components/schemas/Operation'}}}}
...
  /nodes:
    get:
      operationId: listNodes
      parameters: [{$ref: '#/components/parameters/Cursor'}, {$ref: '#/components/parameters/Limit'}, {$ref: '#/components/parameters/Search'}]
      responses:
        '200': {description: Nodes, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/Node'}}}}}
    post:
      operationId: registerNode
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/NodeRegistration'}}}}
      responses:
        '201': {description: Registered, content: {application/json: {schema: {$ref: '#/components/schemas/Node'}}}}
...
components:
  parameters:
    Id: {name: id, in: path, required: true, schema: {type: string}}
    Cursor: {name: cursor, in: query, schema: {type: string}}
    Limit: {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 500, default: 100}}
    Search: {name: search, in: query, schema: {type: string}}
  responses:
    Accepted:
      description: Accepted
      content: {application/json: {schema: {$ref: '#/components/schemas/Operation'}}}
  securitySchemes:
    bearer: {type: http, scheme: bearer}
  schemas:
...
    Resources:
      type: object
      required: [cpu, memoryMiB, diskGiB]
      properties:
        cpu: {type: integer, minimum: 1}
        memoryMiB: {type: integer, format: int64, minimum: 64}
        diskGiB: {type: integer, format: int64, minimum: 1}
...
    Volume:
      type: object
      required: [id, mountPath, sizeGiB]
      properties:
        id: {type: string}
        mountPath: {type: string}
        sizeGiB: {type: integer, format: int64}
        persistentVolumeId: {type: string, description: Registered persistent volume. Its size comes from the volume catalogue.}
...
    NodeVolume:
      type: object
      required: [id, kind, sizeGiB, storage]
      properties:
        id: {type: string}
        kind: {$ref: '#/components/schemas/TemplateKind'}
        sizeGiB: {type: integer, format: int64}
        storage: {$ref: '#/components/schemas/StorageInfo'}
...
    TemplateDisk:
      type: object
      required: [id, sizeGiB, bus, bootOrder]
      properties:
        id: {type: string}
        sizeGiB: {type: integer, format: int64, minimum: 1}
        bus: {type: string, enum: [ide, sata, scsi, virtio]}
        bootOrder: {type: integer, minimum: 1}
        controllerModel: {type: string}
        controllerIndex: {type: integer, minimum: 0}
        controllerUnit: {type: integer, minimum: 0}
...
    TemplateFormat:
      type: string
      enum: [oci, docker, qcow2, raw, vmdk, ova, ovf, iso]
...
    Template:
      type: object
      required: [id, name, kind, os, version, source, resources]
      properties:
        id: {type: string}
        name: {type: string}
        kind: {$ref: '#/components/schemas/TemplateKind'}
        os: {type: string}
        version: {type: integer}
        source: {type: string}
        format: {$ref: '#/components/schemas/TemplateFormat'}
        artifactNodeId: {type: string, readOnly: true}
        operationId: {type: string, readOnly: true}
        resources: {$ref: '#/components/schemas/Resources'}
        hardware: {$ref: '#/components/schemas/Hardware'}
        disks: {type: array, items: {$ref: '#/components/schemas/TemplateDisk'}}
        media: {type: array, items: {$ref: '#/components/schemas/TemplateMedia'}}
        stateFiles: {type: array, items: {type: string}, readOnly: true}
        nicModels: {type: array, items: {type: string}}
        initialization: {$ref: '#/components/schemas/TemplateInitialization'}
        volumes: {type: array, items: {$ref: '#/components/schemas/Volume'}}
        state: {$ref: '#/components/schemas/TemplateState'}
        error: {type: string}
...
    Interface:
      type: object
      required: [id, networkId, mac, address, primary]
      properties:
        id: {type: string}
        networkId: {type: string}
        mac: {type: string}
        address: {type: string}
        primary: {type: boolean}
    Asset:
      type: object
      required: [id, name, templateId, resources, interfaces]
      properties:
        id: {type: string}
        name: {type: string}
        templateId: {type: string}
        storagePoolId: {type: string, description: Omitted selects storage automatically; an existing asset keeps its storage until migrated.}
        resources: {$ref: '#/components/schemas/Resources'}
        interfaces: {type: array, items: {$ref: '#/components/schemas/Interface'}}
        volumes: {type: array, items: {$ref: '#/components/schemas/Volume'}}
        parameters: {type: object, additionalProperties: {type: string}}
        restartPolicy: {$ref: '#/components/schemas/RestartPolicy'}
        guest: {$ref: '#/components/schemas/GuestSettings'}
        media: {type: array, items: {type: string}, description: Attached template media IDs. Omitted attaches all; an empty array ejects all.}
        pciBinding: {$ref: '#/components/schemas/PciBinding'}
    PciBinding:
      type: object
      required: [nodeId, groupIds]
      properties:
        nodeId: {type: string}
        groupIds: {type: array, minItems: 1, uniqueItems: true, items: {type: string}}
    PciGroup:
      type: object
      required: [id, name, available, devices]
      properties:
        id: {type: string}
        name: {type: string}
        available: {type: boolean}
        devices: {type: array, items: {type: string}}
    RestartPolicy:
      type: string
      enum: [never, on-failure, always]
      default: never
      description: Restart unexpected exits only. Explicit stop remains stopped, including across node restarts.
...
    GuestSettings:
      type: object
      properties:
        hostname: {type: string}
        username: {type: string}
        sshAuthorizedKeys: {type: array, items: {type: string}}
...
    TemplateKind:
      type: string
      enum: [container, vm]
...
    Operation:
      type: object
      required: [id, kind, state, phase, completed, total, retryable, createdAt, updatedAt]
      properties:
        id: {type: string}
        environmentId: {type: string}
        kind: {type: string}
        state: {type: string, enum: [queued, running, succeeded, failed, partially_applied]}
        phase: {type: string}
        completed: {type: integer}
        total: {type: integer}
        retryable: {type: boolean}
        error: {type: string}
        results: {type: array, items: {$ref: '#/components/schemas/ExecutionResult'}}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
    Node:
      type: object
      required: [id, name, endpoint, capacity, reserved, capabilities, slots, observedAt]
      properties:
        id: {type: string}
        name: {type: string}
        endpoint: {type: string}
        capacity: {$ref: '#/components/schemas/Resources'}
        reserved: {$ref: '#/components/schemas/Resources'}
        capabilities: {type: array, items: {type: string}}
        vmHardware: {$ref: '#/components/schemas/VmHardware'}
        storageDevice: {type: string, description: Dedicated Ceph disk selected during node installation}
        slots: {type: integer}
        observedAt: {type: string, format: date-time}
        state: {type: string}
        override: {$ref: '#/components/schemas/Resources'}
    ResolvedInterface:
      type: object
      required: [id, networkId, mac, address, prefix, primary, portName, mtu]
      properties:
        id: {type: string}
        networkId: {type: string}
        mac: {type: string}
        address: {type: string}
        prefix: {type: integer}
        primary: {type: boolean}
        portName: {type: string}
        mtu: {type: integer}
        gateway: {type: string}
        dns: {type: array, items: {type: string}}
    AssetExecution:
      type: object
      required: [asset, template, instanceId, interfaces]
      properties:
        asset: {$ref: '#/components/schemas/Asset'}
        template: {$ref: '#/components/schemas/Template'}
        instanceId: {type: string}
        previousInstanceId: {type: string}
        dataSetId: {type: string, x-go-type-skip-optional-pointer: true, description: Actual writable data set; retained across ordinary rebuilds}
        volumeSources: {type: object, additionalProperties: {$ref: '#/components/schemas/NodeVolume'}, x-go-type-skip-optional-pointer: true}
        pciDevices: {type: array, items: {type: string}, x-go-type-skip-optional-pointer: true}
        storagePath: {type: string, description: Node-local managed storage selected by the scheduler}
        storagePoolId: {type: string, description: Actual selected pool; omitted for node default storage}
        storageFilesystem: {type: string, description: Physical filesystem reservation identity}
        rbd: {$ref: '#/components/schemas/RbdStorage'}
        interfaces: {type: array, items: {$ref: '#/components/schemas/ResolvedInterface'}}
...
    TemplateCache:
      type: object
      required: [nodeId, nodeName, bytes, reclaimable, reason]
      properties:
        nodeId: {type: string}
        nodeName: {type: string}
        bytes: {type: integer, format: int64}
        reclaimable: {type: boolean}
        reason: {type: string}
...
    ExecutionResult:
      type: object
      required: [assetId, instanceId, state, observedAt]
      properties:
        environmentId: {type: string}
        nodeId: {type: string}
        assetId: {type: string}
        instanceId: {type: string}
        state: {type: string}
        error: {type: string}
        execution: {$ref: '#/components/schemas/AssetExecution'}
        recovery: {$ref: '#/components/schemas/RecoveryCapture'}
        observedAt: {type: string, format: date-time}
    RecoveryConsistency:
      type: string
      enum: [crash, filesystem, application]
    RecoveryCapture:
      type: object
      required: [sizeBytes, memory, consistency]
      properties:
        sizeBytes: {type: integer, format: int64}
        memory: {type: boolean}
        consistency: {$ref: '#/components/schemas/RecoveryConsistency'}
...
    StorageDriver:
      type: string
      enum: [directory, rbd]
      x-enum-varnames: [StorageDriverDirectory, StorageDriverRBD]
    StorageInfo:
      type: object
      required: [path, filesystem, capacityBytes, availableBytes]
      properties:
        path: {type: string}
        filesystem: {type: string}
        capacityBytes: {type: integer, format: int64}
        availableBytes: {type: integer, format: int64}
        nativeSnapshots: {type: boolean, x-go-type-skip-optional-pointer: true}
        rbd: {$ref: '#/components/schemas/RbdStorage'}
    ConfigureNodeStorage:
      type: object
      required: [device]
      properties:
        device: {type: string, description: Physical empty disk selected for Ceph; empty clears an unused selection}
    StorageDevice:
      type: object
      required: [path, sizeBytes, model, serial, available, reason]
      properties:
        path: {type: string}
        sizeBytes: {type: integer, format: int64}
        model: {type: string}
        serial: {type: string}
        available: {type: boolean}
        reason: {type: string}
    NodeStorageDevices:
      type: object
      required: [devices, selected]
      properties:
        devices: {type: array, items: {$ref: '#/components/schemas/StorageDevice'}}
        selected: {type: string}
        operation: {$ref: '#/components/schemas/Operation'}
    ConfigureCephPool:
      type: object
      required: [replicas]
      properties:
        replicas: {type: integer, minimum: 1, maximum: 3}
    CephDaemon:
      type: object
      required: [name, host, role, state]
      properties:
        name: {type: string}
        host: {type: string}
        role: {type: string}
        state: {type: string}
    CephDisk:
      type: object
      required: [name, host, capacityBytes, usedBytes, state]
      properties:
        name: {type: string}
        host: {type: string}
        capacityBytes: {type: integer, format: int64}
        usedBytes: {type: integer, format: int64}
        state: {type: string}
    CephStatus:
      type: object
      required: [health, messages, replicas, osdsUp, osdsTotal, disks, daemons]
      properties:
        health: {type: string}
        messages: {type: array, items: {type: string}}
        replicas: {type: integer}
        osdsUp: {type: integer}
        osdsTotal: {type: integer}
        disks: {type: array, items: {$ref: '#/components/schemas/CephDisk'}}
        daemons: {type: array, items: {$ref: '#/components/schemas/CephDaemon'}}
    StoragePoolAsset:
      type: object
      required: [environmentId, environmentName, revision, assetId, assetName, nodeId, sizeGiB, state]
      properties:
        environmentId: {type: string}
        environmentName: {type: string}
        revision: {type: integer, format: int64}
        assetId: {type: string}
        assetName: {type: string}
        nodeId: {type: string}
        sizeGiB: {type: integer, format: int64}
        state: {type: string}
    RbdStorage:
      type: object
      required: [pool, imagePrefix, monitors, user, fsid, secretId]
      properties:
        pool: {type: string}
        imagePrefix: {type: string}
        monitors: {type: array, items: {type: string}}
        user: {type: string}
        fsid: {type: string}
        secretId: {type: string}
    CephConnection:
      type: object
      required: [monitors, pool, user, key]
      properties:
        monitors: {type: array, minItems: 1, items: {type: string}}
        pool: {type: string}
        user: {type: string}
        key: {type: string, format: password, writeOnly: true}
    CreateStoragePool:
      type: object
      required: [nodeIds, name, driver]
      properties:
        nodeIds: {type: array, minItems: 1, uniqueItems: true, items: {type: string}}
        name: {type: string}
        driver: {$ref: '#/components/schemas/StorageDriver'}
        directory: {type: string, description: Existing directory or mounted filesystem on this node}
        ceph: {$ref: '#/components/schemas/CephConnection'}
    StoragePool:
      type: object
      required: [id, nodeIds, name, default, driver, allocatedGiB, capabilities]
      properties:
        id: {type: string}
        nodeIds: {type: array, items: {type: string}}
        name: {type: string}
        directory: {type: string, description: Original node storage location; visible to administrators}
        default: {type: boolean}
        driver: {$ref: '#/components/schemas/StorageDriver'}
        state: {type: string, enum: [ready, preparing, deleting], x-enum-varnames: [StoragePoolStateReady, StoragePoolStatePreparing, StoragePoolStateDeleting]}
        managed: {type: boolean, default: false}
        operationState: {type: string}
        operationId: {type: string}
        storage: {$ref: '#/components/schemas/StorageInfo'}
        allocatedGiB: {type: integer, format: int64}
        capabilities: {type: array, items: {type: string}}
        error: {type: string}
...
    VmHardware:
      type: object
      required: [machines, cpuModes, cpuModels, nicModels, diskControllers]
      properties:
        machines: {type: array, items: {$ref: '#/components/schemas/VmMachine'}}
        cpuModes: {type: array, items: {type: string}}
        cpuModels: {type: array, items: {type: string}}
        nicModels: {type: array, items: {type: string}}
        diskControllers: {type: array, items: {type: string}}
        numaPlacement: {type: boolean, description: Host supports libvirt automatic placement through numad.}
        pciGroups: {type: array, items: {$ref: '#/components/schemas/PciGroup'}, x-go-type-skip-optional-pointer: true}
    VmMachine:
      type: object
      required: [name, aliases, maxVcpus, firmware, secureBoot, tpm2, diskBuses, firmwareFiles]
      properties:
        name: {type: string}
        aliases: {type: array, items: {type: string}}
        maxVcpus: {type: integer}
        firmware: {type: array, items: {type: string}}
        secureBoot: {type: boolean}
        tpm2: {type: boolean}
        diskBuses: {type: array, items: {type: string}}
        firmwareFiles: {type: array, items: {type: string}}
...
    NodeRegistration:
      type: object
      required: [name, endpoint]
      properties:
        name: {type: string}
        endpoint: {type: string}
    Problem:
      type: object
      required: [status, title, detail]
      properties:
        status: {type: integer}
        title: {type: string}
        detail: {type: string}
...